requirements.txt
telethon==1.32.1
python-dotenv==1.0.0
loguru==0.7.2 